Update 268. Missing Number
XOR (all numbers in nums array and 0, 1, ..., n)

Each number in [0, 1, ... n] have two occurrences except the missing number. So the result is the missing number. The same idea as the solution to "single number".
i = 0~n-1 through iteration, but we need xor every number 0~n, nums.size() is equal to n make it up